+++ art/art_dic.py
@@ -262,5 +262,3 @@ art_dic = {
 
 
 }
-
-art_counter = len(art_dic)